configuration_api/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
Jinja2==2.11.2
PyYAML==5.3.1
requests==2.25.0
